mac updater: clone prelude if it isn't there

+++ mac_update_things.sh
@@ -25,10 +25,16 @@ git pull
 
 brew update
 
-cd ~/external_src/prelude
-git clean -fdx
-git checkout init.el
-git pull
+if [[ -d ~/external_src/prelude ]]; then
+    cd ~/external_src/prelude
+    git clean -fdx
+    git checkout init.el
+    git pull
+else
+    mkdir -p ~/external_src/prelude
+    git clone https://github.com/bbatsov/prelude ~/external_src/prelude
+fi
+
 rm -rf ~/.emacs.d
 ln -sf ~/external_src/prelude ~/.emacs.d
 cp ~/external_src/prelude/sample/prelude-modules.el ~/.emacs.d/
